requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118
torchvision==0.16.2
torchaudio==2.1.2

torch-geometric==2.4.0
-f https://data.pyg.org/whl/torch-2.1.2+cu118.html

pytorch_lightning==2.2.0
matplotlib==3.6.1
pandas==2.2.2
imageio==2.34.0
wandb==0.16.3
tqdm==4.66.5
pulp==2.8
myopic-mces==1.0
dgllife==0.3.2
hyperopt==0.2.7
numpy==1.24
scikit-learn==1.3.0
